fix refactor <great update of the project window view>
feat: add functionality to modify the enterprise's name and its port's value <done>

feat : serialize the parameter of all pointer we set connection of... <done>

todo : deleting workhours of people we erase from the enterprise

todo: erase excess of data.saveData() in all classes, because we're already saving by closing the stage (by clicking the button 'quit' or the X of the stage)

todo : pointing incident highlighting

todo : synchronize the data from the pointer and the data from the app so that we won't have to refresh the app, to make new workhour appeared after a checkin...
