+++ src/main/java/theView/manage/windowShowEnt/UtilityWindowShowEnt.java
@@ -1,0 +1,25 @@
+package theView.manage.windowShowEnt;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeParseException;
+
+public class UtilityWindowShowEnt {
+    public static boolean isValidTime(String timeStr) {
+        try {
+            LocalTime.parse(timeStr);
+            return true;
+        } catch (DateTimeParseException e) {
+            return false;
+        }
+    }
+
+    public static boolean isValidDate(String dateStr) {
+        try {
+            LocalDate.parse(dateStr);
+            return true;
+        } catch (DateTimeParseException e) {
+            return false;
+        }
+    }
+}
